+++ packages/cron/src/bin/pins-failed.js
@@ -5,12 +5,7 @@ import { fileURLToPath } from 'url'
 import dotenv from 'dotenv'
 import fetch from 'node-fetch'
 import { checkFailedPinStatuses } from '../jobs/pins.js'
-import {
-  getDBClient,
-  getCluster1,
-  getCluster2,
-  getCluster3,
-} from '../lib/utils.js'
+import { getPg, getCluster1, getCluster2, getCluster3 } from '../lib/utils.js'
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
 /** @ts-ignore */
@@ -20,13 +15,21 @@ const oneMonthAgo = () =>
   new Date(new Date().setMonth(new Date().getMonth() - 1))
 
 async function main() {
-  const db = getDBClient(process.env)
-  const cluster1 = getCluster1(process.env)
-  const cluster2 = getCluster2(process.env)
-  const cluster3 = getCluster3(process.env)
-  const after = process.env.AFTER ? new Date(process.env.AFTER) : oneMonthAgo()
+  const pg = getPg(process.env, 'rw')
+  await pg.connect()
 
-  await checkFailedPinStatuses({ db, cluster1, cluster2, cluster3, after })
+  try {
+    const cluster1 = getCluster1(process.env)
+    const cluster2 = getCluster2(process.env)
+    const cluster3 = getCluster3(process.env)
+    const after = process.env.AFTER
+      ? new Date(process.env.AFTER)
+      : oneMonthAgo()
+
+    await checkFailedPinStatuses({ pg, cluster1, cluster2, cluster3, after })
+  } finally {
+    await pg.end()
+  }
 }
 
 dotenv.config({ path: path.join(__dirname, '../../../../.env') })