+++ packages/gateway/test/constants.js
@@ -1,1 +1,5 @@
-export const gateways = ['http://localhost:9081', 'http://localhost:9082']
+export const gateways = [
+  'http://127.0.0.1:9081',
+  'http://localhost:9082',
+  'http://localhost:9083',
+]