+++ packages/client/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+### [5.2.2](https://github.com/nftstorage/nft.storage/compare/client-v5.2.1...client-v5.2.2) (2022-01-24)
+
+
+### Bug Fixes
+
+* **client:** test new release ci ([f26c739](https://github.com/nftstorage/nft.storage/commit/f26c739c39104904152010b8aad5da4914e361c9))
+* **website:** remove type ([94fe5e5](https://github.com/nftstorage/nft.storage/commit/94fe5e50fc581d4ac9062630258934458cb91728))
+
+### [5.2.1](https://www.github.com/nftstorage/nft.storage/compare/nft.storage-v5.2.0...nft.storage-v5.2.1) (2022-01-06)
+
+
+### Bug Fixes
+
+* store directory example ([#1002](https://www.github.com/nftstorage/nft.storage/issues/1002)) ([79cd4bf](https://www.github.com/nftstorage/nft.storage/commit/79cd4bf09d8c8ba66aee4bf3532f790afd4ba8d5))
+
+
+### Changes
+
+* updated dag to ot be restricted ([#1004](https://www.github.com/nftstorage/nft.storage/issues/1004)) ([f002a98](https://www.github.com/nftstorage/nft.storage/commit/f002a98d47db0618ae0d3b48f0d4ef0a12b23da0))
+
 ## [5.2.0](https://www.github.com/nftstorage/nft.storage/compare/nft.storage-v5.1.3...nft.storage-v5.2.0) (2021-12-17)